changing types from double to decimal

+++ NextGen-BM-BE/NextGen-BM-BE-Domain/Entities/PropertyAggregate/Property.cs
@@ -8,9 +8,9 @@ namespace NextGen_BM_BE_Domain.Entities.PropertyAggregate
         public int PropertyId { get; set; }
         public int PropertyNumber { get; set; }
         public int BuildingId { get; set; }
-        public double Size { get; set; }
+        public decimal Size { get; set; }
         public int Floor { get; set; }
-        public double SizeOfIdealParts { get; set; }
+        public decimal SizeOfIdealParts { get; set; }
         public int PropertyTypeId { get; set; }
         public bool EntranceIsExternal { get; set; }
         public ICollection<PropertyExpense>? Expenses { get; set; }
