+++ NextGen-BM-BE/NextGen-BM-BE-Domain/Entities/PropertyAggregate/PropertyUsers.cs
@@ -7,6 +7,6 @@ namespace NextGen_BM_BE_Domain.Entities.PropertyAggregate
         public int UserId { get; set; }
         public DateOnly EffectiveDate { get; set; }
         public DateOnly EndDate { get; set; }
-        public double PercentOfApartmentOwned { get; set; }
+        public decimal PercentOfApartmentOwned { get; set; }
     }
 };
